% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\docType{class}
\name{CMA}
\alias{CMA}
\title{Create a streamer for calculating the cumulative moving average}
\format{
An \code{\link{R6Class}} generator object
}
\description{
\code{CMA} creates a streaming algorithm that can be used to
keep track of the mean of incoming values.
}
\examples{
mean <- CMA$new(c(1, 2))
mean$update(c(3, 4))
mean$value()
#> [1] 2.5


## ------------------------------------------------
## Method `CMA$new`
## ------------------------------------------------

mean <- CMA$new()


## ------------------------------------------------
## Method `CMA$update`
## ------------------------------------------------

mean <- CMA$new()
mean$update(c(1, 2))


## ------------------------------------------------
## Method `CMA$value`
## ------------------------------------------------

mean <- CMA$new(c(1, 2, 3))
mean$value()
#> [1] 2

}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{CMA$new()}}
\item \href{#method-update}{\code{CMA$update()}}
\item \href{#method-value}{\code{CMA$value()}}
\item \href{#method-clone}{\code{CMA$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Creates a new \code{CMA} streamer object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CMA$new(x = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{values to be used during initialisation (optional)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The new \code{CMA} (invisibly)
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{mean <- CMA$new()

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update"></a>}}
\if{latex}{\out{\hypertarget{method-update}{}}}
\subsection{Method \code{update()}}{
Resets the \code{CMA} streamer object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CMA$update(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{values to be added to the stream}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The updated \code{CMA} (invisibly)
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{mean <- CMA$new()
mean$update(c(1, 2))

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-value"></a>}}
\if{latex}{\out{\hypertarget{method-value}{}}}
\subsection{Method \code{value()}}{
Returns the current value of the mean.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CMA$value()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The current value of the \code{CMA} (invisibly)
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{mean <- CMA$new(c(1, 2, 3))
mean$value()
#> [1] 2

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CMA$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
