% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\docType{class}
\name{EMA}
\alias{EMA}
\title{Create a streamer for calculating the exponential moving average}
\format{
An \code{\link{R6Class}} generator object
}
\description{
\code{EMA} creates a streaming algorithm that can be used to
calculate the exponential moving average of incoming values
}
\examples{
exp_mean <- EMA$new(c(1, 2))
exp_mean$update(c(3, 4))
exp_mean$value
#> [1] 3.266667


## ------------------------------------------------
## Method `EMA$new`
## ------------------------------------------------

exp_mean <- EMA$new()


## ------------------------------------------------
## Method `EMA$update`
## ------------------------------------------------

exp_mean <- EMA$new(c(1,2))
exp_mean$update(c(3,4))

}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{EMA$new()}}
\item \href{#method-update}{\code{EMA$update()}}
\item \href{#method-clone}{\code{EMA$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Creates a new \code{EMA} streamer object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EMA$new(x = NULL, alpha = 0.5)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{values to be used during initialization (optional)}

\item{\code{alpha}}{value of the smoothing factor (0.5 by default)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The new \code{EMA} (invisibly)
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{exp_mean <- EMA$new()

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update"></a>}}
\if{latex}{\out{\hypertarget{method-update}{}}}
\subsection{Method \code{update()}}{
Updates the EMA with a stream of new values
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EMA$update(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{a vector of values to update the EMA}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The updated \code{EMA} (invisibly)
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{exp_mean <- EMA$new(c(1,2))
exp_mean$update(c(3,4))

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EMA$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
