% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\docType{class}
\name{Mean}
\alias{Mean}
\title{Create a streamer for calculating a running mean}
\format{
An \code{\link{R6Class}} generator object
}
\description{
\code{Mean} creates a streaming algorithm that can be used to
keep track of the mean of incoming values.
}
\details{
Assemble \code{Node} objects into a \code{data.tree}
structure and use the traversal methods to set, get, and perform operations on it. Typically, you construct larger tree
structures by converting from \code{data.frame}, \code{list}, or other formats.

Most methods (e.g. \code{node$Sort()}) also have a functional form (e.g. \code{Sort(node)})
}
\examples{
mean <- Mean$new(c(1, 2))
mean$update(c(3, 4))
mean$value
#> [1] 2.5


## ------------------------------------------------
## Method `Mean$new`
## ------------------------------------------------

mean <- Mean$new()
mean$reset()



## ------------------------------------------------
## Method `Mean$update`
## ------------------------------------------------

mean <- Mean$new()
mean$update(c(1, 2))


## ------------------------------------------------
## Method `Mean$value`
## ------------------------------------------------

mean <- Mean$new(c(1, 2, 3))
mean$value()
#> [1] 2


## ------------------------------------------------
## Method `Mean$reset`
## ------------------------------------------------

mean <- Mean$new()
mean$reset()

}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Mean$new()}}
\item \href{#method-update}{\code{Mean$update()}}
\item \href{#method-value}{\code{Mean$value()}}
\item \href{#method-reset}{\code{Mean$reset()}}
\item \href{#method-clone}{\code{Mean$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Creates a new \code{Mean} streamer object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Mean$new(x = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{values to be used during initialisation (optional)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The new \code{Mean} (invisibly)
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{mean <- Mean$new()
mean$reset()


}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update"></a>}}
\if{latex}{\out{\hypertarget{method-update}{}}}
\subsection{Method \code{update()}}{
Resets the \code{Mean} streamer object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Mean$update(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{values to be added to the stream}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The updated \code{Mean} (invisibly)
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{mean <- Mean$new()
mean$update(c(1, 2))

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-value"></a>}}
\if{latex}{\out{\hypertarget{method-value}{}}}
\subsection{Method \code{value()}}{
Returns the current value of the mean.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Mean$value()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The updated \code{Mean} (invisibly)
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{mean <- Mean$new(c(1, 2, 3))
mean$value()
#> [1] 2

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-reset"></a>}}
\if{latex}{\out{\hypertarget{method-reset}{}}}
\subsection{Method \code{reset()}}{
Resets the \code{Mean} streamer object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Mean$reset(x = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{values to be used during re-initialisation (optional)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The updated \code{Mean} (invisibly)
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{mean <- Mean$new()
mean$reset()

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Mean$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
