% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\docType{class}
\name{SMA}
\alias{SMA}
\title{Create a streamer for calculating the simple moving average}
\format{
An \code{\link{R6Class}} generator object
}
\description{
\code{SMA} creates a streaming algorithm that can be used to
keep track of the mean of the previous k datapoints
}
\examples{
mean <- SMA$new(c(1, 2, 3, 4, 5), window = 3)
mean$value
#> [1] 4


## ------------------------------------------------
## Method `SMA$new`
## ------------------------------------------------

mean <- SMA$new(window = 5)


## ------------------------------------------------
## Method `SMA$update`
## ------------------------------------------------

mean <- SMA$new(c(1, 2, 3), window = 3)
mean$update(c(4, 5))

}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{SMA$new()}}
\item \href{#method-update}{\code{SMA$update()}}
\item \href{#method-clone}{\code{SMA$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Creates a new \code{SMA} streamer object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SMA$new(x = NULL, window = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{values to be used during initialisation (optional)}

\item{\code{window}}{size of the window}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The new \code{SMA} (invisibly)
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{mean <- SMA$new(window = 5)

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update"></a>}}
\if{latex}{\out{\hypertarget{method-update}{}}}
\subsection{Method \code{update()}}{
Resets the \code{SMA} streamer object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SMA$update(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{values to be added to the stream}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The updated \code{SMA} (invisibly)
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{mean <- SMA$new(c(1, 2, 3), window = 3)
mean$update(c(4, 5))

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SMA$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
