% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\docType{class}
\name{Sorting}
\alias{Sorting}
\title{Create a streamer for sorting incoming data}
\format{
An \code{\link{R6Class}} generator object
}
\description{
\code{Sorting} creates a streaming algorithm that can be used
to sort incoming data
}
\examples{
sorter <- Sorting$new(c(3, 1))
sorter$update(c(2, 4))
sorter$value
#> [1] 1 2 3 4

}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Sorting$new()}}
\item \href{#method-update}{\code{Sorting$update()}}
\item \href{#method-clone}{\code{Sorting$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Creates a new \code{Sorting} streamer object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Sorting$new(x = NULL, method = "linked-list")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{values to be used during initialisation (optional)}

\item{\code{method}}{the method used to sort values}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The new \code{Sorting} (invisibly)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update"></a>}}
\if{latex}{\out{\hypertarget{method-update}{}}}
\subsection{Method \code{update()}}{
Updates the \code{Sorting} streamer object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Sorting$update(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{values to be added to the stream}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The updated \code{Sorting} (invisibly)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Sorting$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
