% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\docType{class}
\name{Variance}
\alias{Variance}
\title{Create a streamer for calculating the population and sample variance}
\format{
An \code{\link{R6Class}} generator object
}
\description{
\code{Variance} creates a streaming algorithm that can be
used to calculate the population and sample variance of incoming values
}
\examples{
variance <- Variance$new(c(1, 2))
variance$update(c(3, 4))
variance$value
#> [1] 1.25


## ------------------------------------------------
## Method `Variance$new`
## ------------------------------------------------

variance <- Variance$new()


## ------------------------------------------------
## Method `Variance$value`
## ------------------------------------------------

variance <- Variance$new(x = c(1,2,3,4))
variance$value()
#> [1] 1.25
variance$value(sample = TRUE)
#> [1] 1.666667


## ------------------------------------------------
## Method `Variance$update`
## ------------------------------------------------

variance <- Variance$new(c(1,2))
variance$update(c(3,4))

}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Variance$new()}}
\item \href{#method-value}{\code{Variance$value()}}
\item \href{#method-update}{\code{Variance$update()}}
\item \href{#method-clone}{\code{Variance$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Creates a new \code{Variance} streamer object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Variance$new(x = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{values to be used during initialization (optional)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The new \code{Variance} (invisibly)
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{variance <- Variance$new()

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-value"></a>}}
\if{latex}{\out{\hypertarget{method-value}{}}}
\subsection{Method \code{value()}}{
Returns the current value of the Variance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Variance$value(sample = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sample}}{for choosing between the population and sample variance.
If `TRUE` the sample variance is returned.
If `FALSE` the population variance is returned.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The current value of the \code{Variance} (invisibly)
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{variance <- Variance$new(x = c(1,2,3,4))
variance$value()
#> [1] 1.25
variance$value(sample = TRUE)
#> [1] 1.666667

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update"></a>}}
\if{latex}{\out{\hypertarget{method-update}{}}}
\subsection{Method \code{update()}}{
Updates the Variance with a stream of new values
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Variance$update(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{a vector of values to update the Variance}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The updated \code{Variance} (invisibly)
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{variance <- Variance$new(c(1,2))
variance$update(c(3,4))

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Variance$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
