% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\docType{class}
\name{WMA}
\alias{WMA}
\title{Create a streamer for calculating the weighted moving average}
\format{
An \code{\link{R6Class}} generator object
}
\description{
\code{WMA} creates a streaming algorithm that can be used to
keep track of the weighted mean of incoming values.

In an n-day WMA the latest day has weight n, the second latest n-1, etc.,
down to one.
}
\examples{
mean <- WMA$new(c(1, 2))
mean$update(c(3, 4))
mean$value()
#> [1] 7.5


## ------------------------------------------------
## Method `WMA$new`
## ------------------------------------------------

weighted_mean <- WMA$new()


## ------------------------------------------------
## Method `WMA$update`
## ------------------------------------------------

weighted_mean <- WMA$new()
weighted_mean$update(c(1, 2))


## ------------------------------------------------
## Method `WMA$value`
## ------------------------------------------------

weighted_mean <- WMA$new(c(1, 2, 3, 4))
weighted_mean$value()
#> [1] 7.5

}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{WMA$new()}}
\item \href{#method-update}{\code{WMA$update()}}
\item \href{#method-value}{\code{WMA$value()}}
\item \href{#method-clone}{\code{WMA$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Creates a new \code{WMA} streamer object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WMA$new(x = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{values to be used during initialisation (optional)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The new \code{WMA} (invisibly)
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{weighted_mean <- WMA$new()

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update"></a>}}
\if{latex}{\out{\hypertarget{method-update}{}}}
\subsection{Method \code{update()}}{
Resets the \code{WMA} streamer object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WMA$update(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{values to be added to the stream}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The updated \code{WMA} (invisibly)
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{weighted_mean <- WMA$new()
weighted_mean$update(c(1, 2))

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-value"></a>}}
\if{latex}{\out{\hypertarget{method-value}{}}}
\subsection{Method \code{value()}}{
Returns the current value of the weighted average.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WMA$value()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The current value of the \code{WMA} (invisibly)
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{weighted_mean <- WMA$new(c(1, 2, 3, 4))
weighted_mean$value()
#> [1] 7.5

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WMA$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
