% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machine_learning.R
\docType{class}
\name{LinearRegression}
\alias{LinearRegression}
\title{Create a streamer for updating a linear regression model}
\format{
An \code{\link{R6Class}} generator object
}
\description{
\code{LinearRegressions} creates a streaming algorithm that can
calculate and update linear regression coefficients.
}
\examples{
X <- matrix(c(3, 1, 4, 1, 5, 9, 2, 6), ncol = 2)
y <- c(2, 7, 1, 8)
reg <- LinearRegression$new(X, y)
X_new <- matrix(c(5, 3, 5, 8), ncol = 2)
y_new <- c(2, 8)
reg$update(X_new, y_new)

}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{LinearRegression$new()}}
\item \href{#method-update}{\code{LinearRegression$update()}}
\item \href{#method-value}{\code{LinearRegression$value()}}
\item \href{#method-clone}{\code{LinearRegression$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Creates a new \code{LinearRegression} streamer object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LinearRegression$new(X, y, fit_intercept = TRUE, method = "woodbury")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{X}}{initial design matrix}

\item{\code{y}}{initial labels}

\item{\code{fit_intercept}}{if `TRUE`, add an intercept to the model}

\item{\code{method}}{the method used to update regression coefficients}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The new \code{LinearRegression} (invisibly)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update"></a>}}
\if{latex}{\out{\hypertarget{method-update}{}}}
\subsection{Method \code{update()}}{
Updates the regression coefficient based on new data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LinearRegression$update(X, y)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{X}}{new design matrix}

\item{\code{y}}{new labels

#' @return The updated \code{LinearRegression} (invisibly)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-value"></a>}}
\if{latex}{\out{\hypertarget{method-value}{}}}
\subsection{Method \code{value()}}{
Returns the current coefficient vector.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LinearRegression$value()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The current coefficient vector
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LinearRegression$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
